requirements.txt
torch
torchvision
numpy
pandas
opencv-python
scikit-learn
Pillow
gdown
scikit-image
